requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
anthropic==0.49.0
httpx==0.28.1
pydantic==2.10.4
pydantic-settings==2.7.1
sse-starlette==2.2.1
aiosqlite==0.20.0
PyGithub==2.5.0
python-dotenv==1.0.1
cryptography>=42.0.0
scikit-learn>=1.5.0
joblib>=1.4.0
numpy>=1.26.0
